move SingleValueListAttribute to manager
